+++ docs/changes.md
@@ -24,6 +24,7 @@ Changes in command line arguments:
 - [NXDRIVE-1241](https://jira.nuxeo.com/browse/NXDRIVE-1241): Set the --consider-ssl-errors argument default value to True
 - [NXDRIVE-1242](https://jira.nuxeo.com/browse/NXDRIVE-1242): Use type annotations everywhere
 - [NXDRIVE-1256](https://jira.nuxeo.com/browse/NXDRIVE-1256): Fix the missing synchronization folder icon on Windows
+- [NXDRIVE-1258](https://jira.nuxeo.com/browse/NXDRIVE-1258): Better handle unknown digest algorithms
 - [NXDRIVE-1262](https://jira.nuxeo.com/browse/NXDRIVE-1262): Do not rely on XCode on macOS
 - [NXDRIVE-1279](https://jira.nuxeo.com/browse/NXDRIVE-1279): Use flake8 for another clean-up round
 - [NXDRIVE-1286](https://jira.nuxeo.com/browse/NXDRIVE-1286): Ignore documents when the system cannot find the file
@@ -67,7 +68,7 @@ Changes in command line arguments:
 - Jenkins: Removed the `ENABLE_PROFILER` parameter from the Drive-tests job
 - Packaging: Added `distro` 1.3.0
 - Packaging: Added `flake8` 3.5.0
-- Packaging: Added `nuxeo` 2.0.1
+- Packaging: Added `nuxeo` 2.0.3
 - Packaging: Added `pre-commit` 1.11.0
 - Packaging: Updated `psutil` from 5.4.4 to 5.4.7
 - Packaging: Updated `pycryptodomex` from 3.6.1 to 3.6.6